Improved Test-FMU generation script.

+++ TestFMUs/generate_FMUs.sh
@@ -2,60 +2,63 @@
 
 # create subdir and copy all files together
 
-mkdir fmus
+mkdir -p fmus
 cd fmus
 
 # --- Part 1 ---
 
-mkdir Part1
+mkdir -p Part1/binaries/linux64
 cd Part1
 
-mkdir binaries
-mkdir binaries/linux64
-cp ../../../bin/release/libMath003Part1.so binaries/linux64
+cp ../../../bin/release/libMath003Part1.so binaries/linux64/Part1.so
 cp ../../Math003Part1/data/modelDescription.xml .
 
-rm ../Part1.fmu
+if [ -f ../Part1.fmu ]; then
+    rm ../Part1.fmu
+fi
 
-7za a ../Part1.zip .
-cd ..
-mv Part1.zip Part1.fmu
+7za a ../Part1.zip . > nul && 
+cd .. && 
+mv Part1.zip Part1.fmu &&
+echo "Created Part1.fmu"
 
 
 # --- Part 2 ---
 
 
-mkdir Part2
+mkdir -p Part2/binaries/linux64
 cd Part2
 
-mkdir binaries
-mkdir binaries/linux64
-cp ../../../bin/release/libMath003Part2.so binaries/linux64
+cp ../../../bin/release/libMath003Part2.so binaries/linux64/Part1.so
 cp ../../Math003Part2/data/modelDescription.xml .
 
-rm ../Part2.fmu
+if [ -f ../Part2.fmu ]; then
+    rm ../Part2.fmu
+fi
 
-7za a ../Part2.zip .
-cd ..
-mv Part2.zip Part2.fmu
+7za a ../Part2.zip . > nul &&
+cd .. &&
+mv Part2.zip Part2.fmu &&
+echo "Created Part2.fmu"
 
 
 # --- Part 3 ---
 
 
-mkdir Part3
+mkdir -p Part3/binaries/linux64
 cd Part3
 
-mkdir binaries
-mkdir binaries/linux64
-cp ../../../bin/release/libMath003Part3.so binaries/linux64
+cp ../../../bin/release/libMath003Part3.so binaries/linux64/Part1.so
 cp ../../Math003Part3/data/modelDescription.xml .
 
-rm ../Part3.fmu
+if [ -f ../Part3.fmu ]; then
+    rm ../Part3.fmu
+fi
 
-7za a ../Part3.zip .
-cd ..
-mv Part3.zip Part3.fmu
+7za a ../Part3.zip . > nul &&
+cd .. &&
+mv Part3.zip Part3.fmu &&
+echo "Created Part3.fmu"
 
 
 # go back to original directory
